move singleplayer tossups and bonuses to their own folder
